requirements.txt
pytaglib==3.0.0
PyYAML==6.0.2
